pyproject.toml
[tool.poetry]
name = "nonebot-plugin-phlogo"
version = "0.0.2"
description = "生成ph风格logo"
authors = ["kexue <x@kexue.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.3,<3.10"
Pillow = "^8.3.1"
nonebot-adapter-onebot = "2.0.0-beta.1"
nonebot2 = "2.0.0-beta.1"

[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
